Fix bug
 * fixed a bug where effective status of the updated test result was always 'PASS'
 * changed id data type from Int to Long for all JIRA DTOs
 * added tests
 * added documentation

+++ src/main/kotlin/com/github/jokoroukwu/zephyrng/http/AsyncZephyrPublisher.kt
@@ -40,9 +40,9 @@ class AsyncZephyrPublisher(
      * @param testResultStatusToIdMap a map where every each Zephyr status is mapped to the corresponding id
      */
     fun submitResults(
-        projectId: Int,
+        projectId: Long,
         testNgZephyrSuites: Collection<TestNgZephyrSuite>,
-        testResultStatusToIdMap: Map<TestResultStatus, Int>
+        testResultStatusToIdMap: Map<TestResultStatus, Long>
     ) {
         if (testNgZephyrSuites.isNotEmpty()) {
             logger.info { "Publishing TestNG test results" }
