+++ src/main/kotlin/com/github/jokoroukwu/zephyrng/http/detailedreport/ZephyrStepResult.kt
@@ -3,4 +3,4 @@ package com.github.jokoroukwu.zephyrng.http.detailedreport
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ZephyrStepResult(val id: Int, val index: Int)
+data class ZephyrStepResult(val id: Long, val index: Int)
