+++ src/main/kotlin/com/github/jokoroukwu/zephyrng/http/gettestcases/ZephyrProjectWithTestCases.kt
@@ -1,6 +1,6 @@
 package com.github.jokoroukwu.zephyrng.http.gettestcases
 
 data class ZephyrProjectWithTestCases(
-    val projectId: Int,
-    val testCaseKeyToIdMap: Map<String, Int>
+    val projectId: Long,
+    val testCaseKeyToIdMap: Map<String, Long>
 )
