+++ src/main/kotlin/com/github/jokoroukwu/zephyrng/http/testresultstatus/SerializableTestResultStatusItem.kt
@@ -3,4 +3,4 @@ package com.github.jokoroukwu.zephyrng.http.testresultstatus
 import kotlinx.serialization.Serializable
 
 @Serializable
-class SerializableTestResultStatusItem(val id: Int, val name: TestResultStatus)
+class SerializableTestResultStatusItem(val id: Long, val name: TestResultStatus)
