+++ src/main/kotlin/com/github/jokoroukwu/zephyrng/suitetestcaseidmerger/TestCaseIdInjector.kt
@@ -6,6 +6,6 @@ interface TestCaseIdInjector {
 
     fun injectTestCaseIds(
         testSuitesWithResults: List<TestNgZephyrSuite>,
-        testCaseKeyToIdMap: Map<String, Int>
+        testCaseKeyToIdMap: Map<String, Long>
     ): TestCasesIdInjectResult
 }
