+++ src/main/kotlin/com/github/jokoroukwu/zephyrng/updatableresult/merge/ResultMergeStrategy.kt
@@ -6,7 +6,7 @@ import com.github.jokoroukwu.zephyrng.http.testresultstatus.TestResultStatus
 interface ResultMergeStrategy {
 
     fun mergeResults(
-        testResultStatusToIdMap: Map<TestResultStatus, Int>,
+        testResultStatusToIdMap: Map<TestResultStatus, Long>,
         zephyrDataSet: List<ZephyrStepResult>,
         testNgDataSetResult: com.github.jokoroukwu.zephyrng.TestNgDataSetResult
     ): MergeResult
